Add Project Euler Problem 034 solution
Added a new project, "Problem 034," to the Project Euler solution file (`Project Euler.sln`). This includes project configuration and build settings for Win32 and x64 platforms in both Debug and Release configurations.

Created `Main.cpp` for "Problem 034" with the solution to find the sum of all numbers equal to the sum of the factorial of their digits.

Added `Problem 034.vcxproj` and `Problem 034.vcxproj.filters` to define project settings and organize files.