Add Project Euler Problem 036 solution
Added a new project "Problem 036" to the Visual Studio solution file `Project Euler.sln` with configurations for x86 and x64 platforms in both Debug and Release modes. Updated the `GlobalSection(SolutionConfigurationPlatforms)` in `Project Euler.sln` to include the new project's build configurations.

Created `Main.cpp` for "Problem 036" containing the solution to find the sum of all numbers less than one million that are palindromic in both base 10 and base 2.

Added `Problem 036.vcxproj` and `Problem 036.vcxproj.filters` to configure and organize the project files in Visual Studio.