Add Project 041: Pandigital Prime solution
This commit introduces a new project for "Problem 041" in the Project Euler solution. It includes a new source file `Main.cpp` that implements the logic to find the largest n-digit pandigital prime. Additionally, project files (`Problem 041.vcxproj` and `Problem 041.vcxproj.filters`) are created to define project settings and configurations. The solution file (`Project Euler.sln`) is updated to include the new project and its build configurations.