Add Project 046: Goldbach's Other Conjecture solution
This commit introduces a new project for "Problem 046" in the solution, featuring a C++ implementation that explores Goldbach's Other Conjecture. The project includes `.vcxproj` and `.vcxproj.filters` files, and implements functions for generating prime numbers using the Sieve of Eratosthenes, as well as checking if odd composite numbers can be expressed as the sum of a prime and twice a square. It is configured for both Debug and Release modes for Win32 and x64 platforms.