Add Problem 050 project to Project Euler solution
This commit introduces a new project "Problem 050" to the "Project Euler.sln" solution. It includes the project files "Problem 050.vcxproj" and "Problem 050.vcxproj.filters", along with a C++ source file "Main.cpp". The implementation solves Project Euler's Problem 050 by finding the prime that can be expressed as the sum of the most consecutive primes below one million. The project supports both Win32 and x64 configurations in Debug and Release modes.